Renaming file GOSettingsMidiMatchDialog to GOSettingsDeviceMatchDialog
