+++ src/grandorgue/dialogs/settings/GOSettingsMidiDeviceList.cpp
@@ -7,8 +7,6 @@
 
 #include "GOSettingsMidiDeviceList.h"
 
-#include "config/GOMidiDeviceConfig.h"
-
 #include "GOSettingsDeviceMatchDialog.h"
 
 GOSettingsMidiDeviceList::GOSettingsMidiDeviceList(
